# planner/core: thoroughly push down count-distinct agg in the MPP mode. (#25662)