--- a/planner/core/task.go
+++ b/planner/core/task.go
@@ -1926,9 +1926,9 @@ func (p *PhysicalHashAgg) attach2TaskForMpp(tasks ...task) task {
 		}
 		// TODO: how to set 2-phase cost?
 		newMpp.addCost(p.GetCost(inputRows, false, true))
-		finalAgg.SetCost(mpp.cost())
+		finalAgg.SetCost(newMpp.cost())
 		if proj != nil {
-			proj.SetCost(mpp.cost())
+			proj.SetCost(newMpp.cost())
 		}
 		return newMpp
 	case MppTiDB:
@@ -1946,6 +1946,25 @@ func (p *PhysicalHashAgg) attach2TaskForMpp(tasks ...task) task {
 		t.addCost(p.GetCost(inputRows, true, false))
 		finalAgg.SetCost(t.cost())
 		return t
+	case MppScalar:
+		proj := p.convertAvgForMPP()
+		partialAgg, finalAgg := p.newPartialAggregate(kv.TiFlash, true)
+		if partialAgg == nil || finalAgg == nil {
+			return invalidTask
+		}
+		attachPlan2Task(partialAgg, mpp)
+		prop := &property.PhysicalProperty{TaskTp: property.MppTaskType, ExpectedCnt: math.MaxFloat64, MPPPartitionTp: property.AnyType}
+		newMpp := mpp.enforceExchangerImpl(prop)
+		attachPlan2Task(finalAgg, newMpp)
+		if proj != nil {
+			attachPlan2Task(proj, newMpp)
+		}
+		newMpp.addCost(p.GetCost(inputRows, false, true))
+		finalAgg.SetCost(newMpp.cost())
+		if proj != nil {
+			proj.SetCost(newMpp.cost())
+		}
+		return newMpp
 	default:
 		return invalidTask
 	}