--- a/planner/core/task.go
+++ b/planner/core/task.go
@@ -1956,14 +1956,19 @@ func (p *PhysicalHashAgg) attach2TaskForMpp(tasks ...task) task {
 		prop := &property.PhysicalProperty{TaskTp: property.MppTaskType, ExpectedCnt: math.MaxFloat64, MPPPartitionTp: property.AnyType}
 		newMpp := mpp.enforceExchangerImpl(prop)
 		attachPlan2Task(finalAgg, newMpp)
-		if proj != nil {
-			attachPlan2Task(proj, newMpp)
+		if proj == nil {
+			proj = PhysicalProjection{
+				Exprs: make([]expression.Expression, 0, len(p.Schema().Columns)),
+			}.Init(p.ctx, p.statsInfo(), p.SelectBlockOffset())
+			for _, col := range p.Schema().Columns {
+				proj.Exprs = append(proj.Exprs, col)
+			}
+			proj.SetSchema(p.schema)
 		}
+		attachPlan2Task(proj, newMpp)
 		newMpp.addCost(p.GetCost(inputRows, false, true))
 		finalAgg.SetCost(newMpp.cost())
-		if proj != nil {
-			proj.SetCost(newMpp.cost())
-		}
+		proj.SetCost(newMpp.cost())
 		return newMpp
 	default:
 		return invalidTask